openclaw-enterprise/scripts/requirements-runtime.txt
fastapi>=0.115
uvicorn>=0.30
prometheus-client>=0.20
psycopg2-binary>=2.9
pyyaml>=6.0
